requirements.txt
asgiref==3.4.1
autopep8==1.5.7
Django==3.2.7
django-filter==21.1
djangorestframework==3.12.4
Markdown==3.3.4
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
